refactor(mcp): extract analysis tools to tools_analysis.py
Move 6 analysis tools (compare_sessions, firebase_audit, analyze_js_bundles,
discover_api, reason_chains, discover_services) from tools.py into a new
tools_analysis.py module with register_analysis_tools(). Pure refactor with
no behavior changes. Removes unused imports from tools.py.
